requirements.txt
Django>=5.2.5
psycopg2-binary>=2.9.0  # For PostgreSQL support
python-decouple>=3.8  # For environment variables
Pillow>=10.0.0  # For image handling if needed
reportlab>=4.0.0  # For PDF generation
gunicorn>=21.2.0  # Production server
whitenoise>=6.5.0  # Serving static files
django-storages>=1.14  # For Azure Storage
azure-storage-blob>=12.17.0  # For Azure Blob Storage
python-dotenv>=1.0.0  # For environment variables